Build/better gen (#132)
* build: directly generate js libs to node_modules

Roll back https://github.com/stellar/soroban-example-dapp/pull/122, but
keep `crowdfund-contract` and `abundance-token` in
`optionalDependencies`.

As I worked with the old version in earnest, iterating on a local copy
of the CLI and generating new JS libs over and over, the two-step
generate-to-.soroban, then install-to-node_modules wasn't working very
well. Probably because the new library wasn't being added until the
`postinstall` step, and the dependencies are optional.

Generating directly to `node_modules` feels dirty, but it works
consistently. At least in this version, you can still see the
dependencies listed in the `optionalDependencies` section, so they're
not quite as surprising/mysterious.

* build: try adding back optionalDeps

* build: update node; use preinstall and local deps

* Update package-lock.json

* work around a bug in the binding gen

---------

Co-authored-by: Chad Ostrowski <221614+chadoh@users.noreply.github.com>